BadDesignFullProject

+++ OnlineShop.Models/Models/ProductModel.cs
@@ -6,7 +6,7 @@ namespace OnlineShop.Models
     {
         public int Id { get; set; }
         public string ProductName { get; set; }
-        public double ProductPrice { get; set; }
+        public decimal ProductPrice { get; set; }
         public string ProductPhotoURL { get; set; }
     }
 }
